requirements.txt
boto3==1.26.137
botocore==1.29.137
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
google-api-core==2.11.0
google-auth==2.19.0
google-cloud-language==2.9.1
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.54.2
idna==3.4
jmespath==1.0.1
proto-plus==1.22.2
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
requests==2.31.0
rsa==4.9
s3transfer==0.6.1
six==1.16.0
tomli==2.0.1
urllib3==1.26.15
yapf==0.33.0
